azure-drift/requirements.txt
azure-identity==1.15.0
azure-mgmt-monitor==5.0.0
azure-mgmt-resource==23.0.1
azure-mgmt-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-keyvault==10.2.0
azure-mgmt-compute==30.3.0
azure-mgmt-network==25.1.0
azure-mgmt-storage==20.0.0
azure-mgmt-sql==1.0.1
azure-mgmt-web==2.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-containerservice==26.1.0
azure-mgmt-cosmosdb==7.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-redis==14.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-appservice==2.0.0
azure-mgmt-loganalytics==1.0.0
azure-mgmt-security==2.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-frontdoor==1.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-policyinsights==1.0.0
azure-mgmt-reservations==1.0.0
azure-mgmt-billing==1.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-consumption==10.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-labservices==1.0.0
azure-mgmt-automation==1.0.0
azure-mgmt-batch==16.0.0
azure-mgmt-cognitiveservices==13.3.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-databricks==1.0.0
azure-mgmt-datafactory==2.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-eventgrid==10.0.0
azure-mgmt-eventhub==10.0.0
azure-mgmt-hdinsight==9.0.0
azure-mgmt-iotcentral==10.0.0
azure-mgmt-iotspaces==0.2.0
azure-mgmt-logic==10.0.0
azure-mgmt-machinelearningservices==1.0.0
azure-mgmt-media==9.0.0
azure-mgmt-mixedreality==1.0.0
azure-mgmt-netapp==1.0.0
azure-mgmt-notificationhubs==8.0.0
azure-mgmt-powerbiembedded==1.0.0
azure-mgmt-recoveryservices==2.0.0
azure-mgmt-recoveryservicesbackup==7.0.0
azure-mgmt-relay==1.0.0
azure-mgmt-search==8.0.0
azure-mgmt-servicebus==8.0.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-signalr==1.0.0
azure-mgmt-sqlvirtualmachine==1.0.0
azure-mgmt-storagecache==1.0.0
azure-mgmt-storagesync==1.0.0
azure-mgmt-subscription==1.0.0
azure-mgmt-timeseriesinsights==1.0.0
azure-mgmt-vmwarecloudsimple==1.0.0
azure-mgmt-webpubsub==1.0.0
azure-mgmt-workloadmonitor==1.0.0
msal==1.25.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.3
numpy==1.26.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Testing dependencies
pytest-mock==3.12.0
httpx==0.25.1
coverage==7.3.2

# Observability
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-logging==0.41b0
prometheus-client==0.17.1
structlog==23.1.0 